+++ moment/core_test.go
@@ -58,7 +58,7 @@ func TestNewCore(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		assert.Equal(t, uint64(3), core.window)
+		assert.Equal(t, 3, core.window)
 		assert.Equal(t, make([]float64, 6), core.sums)
 		assert.Equal(t, uint64(0), core.queue.Len())
 	})
